finish gRPC refactor with nodeIntegration off
