+++ main.js
@@ -553,7 +553,6 @@ ipcMain.on("fetch-meta-and-client", (event, data) => {
   const { rpcType } = data;
   const PROTO_PATH = reqResObj.protoPath;
   const { packageName, service, url, rpc, queryArr } = data.reqResObj;
-
   const packageDefinition = protoLoader.loadSync(PROTO_PATH, {
     keepCase: true,
     longs: String,
@@ -612,7 +611,44 @@ ipcMain.on("fetch-meta-and-client", (event, data) => {
         });
         mainWindow.webContents.send("reqResUpdate", reqResObj);
       });
+  } else if (rpcType === "SERVER STREAM") {
+    console.log("SERVER STREAM inside gRPC");
+    const timesArr = [];
+    // Open Connection for SERVER Stream
+    reqResObj.connection = "open";
+    reqResObj.timeSent = Date.now();
+    const call = client[rpc](reqResObj.queryArr[0], meta);
+    call.on("data", (resp) => {
+      const time = {};
+      time.timeReceived = Date.now();
+      time.timeSent = reqResObj.timeSent;
+      // add server response to reqResObj and dispatch to state/store
+      reqResObj.response.events.push(resp);
+      reqResObj.response.times.push(time);
+      reqResObj.timeReceived = time.timeReceived; //  overwritten on each call to get the final value
+
+      mainWindow.webContents.send("reqResUpdate", reqResObj);
+    });
+    call.on("error", () => {
+      // for fatal error from server
+      console.log("server side stream error");
+    });
+    call.on("end", () => {
+      // Close Connection for SERVER Stream
+      reqResObj.connection = "closed";
+      // no need to push response to reqResObj, no event expected from on 'end'
+      mainWindow.webContents.send("reqResUpdate", reqResObj);
+    });
+    call.on("metadata", (data) => {
+      const metadata = data.internalRepr;
+      // set metadata Map as headers
+      metadata.forEach((value, key) => {
+        reqResObj.response.headers[key] = value[0];
+      });
+      mainWindow.webContents.send("reqResUpdate", reqResObj);
+    });
   } else if (rpcType === "CLIENT STREAM") {
+    console.log("CLIENT STREAM inside gRPC");
     // create call and open client stream connection
     reqResObj.connection = "open";
     const timeSent = Date.now();
@@ -660,93 +696,59 @@ ipcMain.on("fetch-meta-and-client", (event, data) => {
     }
     call.end();
   }
-  // else if (rpcType === "SERVER STREAM") {
-  //   const timesArr = [];
-  //   // Open Connection for SERVER Stream
-  //   reqResObj.connection = "open";
-  //   reqResObj.timeSent = Date.now();
-  //   const call = client[rpc](reqResObj.queryArr[0], meta);
-  //   call.on("data", (resp) => {
-  //     const time = {};
-  //     time.timeReceived = Date.now();
-  //     time.timeSent = reqResObj.timeSent;
-  //     // add server response to reqResObj and dispatch to state/store
-  //     reqResObj.response.events.push(resp);
-  //     reqResObj.response.times.push(time);
-  //     reqResObj.timeReceived = time.timeReceived; //  overwritten on each call to get the final value
 
-  //     store.default.dispatch(actions.reqResUpdate(reqResObj));
-  //   });
-  //   call.on("error", () => {
-  //     // for fatal error from server
-  //     console.log("server side stream erring out");
-  //   });
-  //   call.on("end", () => {
-  //     // Close Connection for SERVER Stream
-  //     reqResObj.connection = "closed";
-  //     // no need to push response to reqResObj, no event expected from on 'end'
-  //     store.default.dispatch(actions.reqResUpdate(reqResObj));
-  //   });
-  //   call.on("metadata", (metadata) => {
-  //     const keys = Object.keys(metadata._internal_repr);
-  //     for (let i = 0; i < keys.length; i += 1) {
-  //       const key = keys[i];
-  //       reqResObj.response.headers[key] = metadata._internal_repr[key][0];
-  //     }
-  //     store.default.dispatch(actions.reqResUpdate(reqResObj));
-  //   });
-  // }
-  // //else BIDIRECTIONAL
-  // else {
-  //   // Open duplex stream
-  //   let counter = 0;
-  //   const call = client[rpc](meta);
-  //   call.on("data", (response) => {
-  //     const curTimeObj = reqResObj.response.times[counter];
-  //     counter++;
-  //     //Close Individual Server Response for BIDIRECTIONAL Stream
-  //     reqResObj.connection = "pending";
-  //     curTimeObj.timeReceived = Date.now();
-  //     reqResObj.timeReceived = curTimeObj.timeReceived;
-  //     reqResObj.response.events.push(response);
-  //     reqResObj.response.times.push(curTimeObj);
-  //     store.default.dispatch(actions.reqResUpdate(reqResObj));
-  //   }); // metadata from server
-  //   call.on("metadata", (metadata) => {
-  //     const keys = Object.keys(metadata._internal_repr);
-  //     for (let i = 0; i < keys.length; i += 1) {
-  //       const key = keys[i];
-  //       reqResObj.response.headers[key] = metadata._internal_repr[key][0];
-  //     }
-  //     store.default.dispatch(actions.reqResUpdate(reqResObj));
-  //   });
-  //   call.on("error", () => {
-  //     console.log("server ended connection with error");
-  //   });
-  //   call.on("end", (data) => {
-  //     //Close Final Server Connection for BIDIRECTIONAL Stream
-  //     reqResObj.connection = "closed";
-  //     // no need to push response to reqResObj, no event expected from on 'end'
-  //     store.default.dispatch(actions.reqResUpdate(reqResObj));
-  //   });
+  //else BIDIRECTIONAL
+  else {
+    console.log("BIDIRECTIONAL gRPC");
+    let counter = 0;
+    const call = client[rpc](meta);
+    call.on("data", (response) => {
+      const curTimeObj = reqResObj.response.times[counter];
+      counter++;
+      //Close Individual Server Response for BIDIRECTIONAL Stream
+      reqResObj.connection = "pending";
+      curTimeObj.timeReceived = Date.now();
+      reqResObj.timeReceived = curTimeObj.timeReceived;
+      reqResObj.response.events.push(response);
+      reqResObj.response.times.push(curTimeObj);
+      // update redux store
+      mainWindow.webContents.send("reqResUpdate", reqResObj);
+    }); // metadata from server
+    call.on("metadata", (data) => {
+      const metadata = data.internalRepr;
 
-  //   for (let i = 0; i < queryArr.length; i++) {
-  //     const time = {};
-  //     const query = queryArr[i];
-  //     //Open Connection for BIDIRECTIONAL Stream
-  //     if (i === 0) {
-  //       reqResObj.connection = "open";
-  //     } else {
-  //       reqResObj.connection = "pending";
-  //     }
-  //     time.timeSent = Date.now();
-  //     reqResObj.timeSent = time.timeSent;
-  //     reqResObj.response.times.push(time);
-  //     call.write(query);
-  //   }
-  //   call.end();
-  // }
-  // store.default.dispatch(actions.reqResUpdate(reqResObj));
+      metadata.forEach((value, key) => {
+        reqResObj.response.headers[key] = value[0];
+      });
+      mainWindow.webContents.send("reqResUpdate", reqResObj);
+    });
+    call.on("error", () => {
+      console.log("server ended connection with error");
+    });
+    call.on("end", (data) => {
+      //Close Final Server Connection for BIDIRECTIONAL Stream
+      reqResObj.connection = "closed";
+      // no need to push response to reqResObj, no event expected from on 'end'
+      mainWindow.webContents.send("reqResUpdate", reqResObj);
+    });
+
+    for (let i = 0; i < queryArr.length; i++) {
+      const time = {};
+      const query = queryArr[i];
+      //Open Connection for BIDIRECTIONAL Stream
+      if (i === 0) {
+        reqResObj.connection = "open";
+      } else {
+        reqResObj.connection = "pending";
+      }
+      time.timeSent = Date.now();
+      reqResObj.timeSent = time.timeSent;
+      reqResObj.response.times.push(time);
+      call.write(query);
+    }
+    call.end();
+  }
+  mainWindow.webContents.send("reqResUpdate", reqResObj);
 });
 
 // ====================== OLDER STUFF =======================
